requirements.txt
openai>=1.0.0
discord.py
requests
python-dotenv
